将id字段修改为Long

+++ musicapi/src/main/java/com/cyl/musicapi/netease/LoginInfo.kt
@@ -6,41 +6,41 @@ import com.google.gson.annotations.SerializedName
 data class Account(@SerializedName("salt")
                    val salt: String = "",
                    @SerializedName("vipType")
-                   val vipType: Int = 0,
+                   val vipType: Long = 0,
                    @SerializedName("userName")
                    val userName: String = "",
                    @SerializedName("type")
-                   val type: Int = 0,
+                   val type: Long = 0,
                    @SerializedName("ban")
-                   val ban: Int = 0,
+                   val ban: Long = 0,
                    @SerializedName("anonimousUser")
                    val anonimousUser: Boolean = false,
                    @SerializedName("createTime")
                    val createTime: Long = 0,
                    @SerializedName("tokenVersion")
-                   val tokenVersion: Int = 0,
+                   val tokenVersion: Long = 0,
                    @SerializedName("id")
-                   val id: Int = 0,
+                   val id: Long = 0,
                    @SerializedName("whitelistAuthority")
-                   val whitelistAuthority: Int = 0,
+                   val whitelistAuthority: Long = 0,
                    @SerializedName("baoyueVersion")
-                   val baoyueVersion: Int = 0,
+                   val baoyueVersion: Long = 0,
                    @SerializedName("viptypeVersion")
-                   val viptypeVersion: Int = 0,
+                   val viptypeVersion: Long = 0,
                    @SerializedName("donateVersion")
-                   val donateVersion: Int = 0,
+                   val donateVersion: Long = 0,
                    @SerializedName("status")
-                   val status: Int = 0)
+                   val status: Long = 0)
 
 
 data class LoginInfo(@SerializedName("clientId")
                      val clientId: String = "",
                      @SerializedName("code")
-                     val code: Int = 0,
+                     val code: Long = 0,
                      @SerializedName("msg")
                      val msg: String = "",
                      @SerializedName("loginType")
-                     val loginType: Int = 0,
+                     val loginType: Long = 0,
                      @SerializedName("profile")
                      val profile: Profile,
                      @SerializedName("bindings")
@@ -52,7 +52,7 @@ data class LoginInfo(@SerializedName("clientId")
 
 
 data class BindingsItem(@SerializedName("expiresIn")
-                        val expiresIn: Int = 0,
+                        val expiresIn: Long = 0,
                         @SerializedName("expired")
                         val expired: Boolean = false,
                         @SerializedName("tokenJsonStr")
@@ -62,9 +62,9 @@ data class BindingsItem(@SerializedName("expiresIn")
                         @SerializedName("id")
                         val id: Long = 0,
                         @SerializedName("type")
-                        val type: Int = 0,
+                        val type: Long = 0,
                         @SerializedName("userId")
-                        val userId: Int = 0,
+                        val userId: Long = 0,
                         @SerializedName("url")
                         val url: String = "")
 
@@ -76,9 +76,9 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("backgroundUrl")
                    val backgroundUrl: String = "",
                    @SerializedName("gender")
-                   val gender: Int = 0,
+                   val gender: Long = 0,
                    @SerializedName("city")
-                   val city: Int = 0,
+                   val city: Long = 0,
                    @SerializedName("signature")
                    val signature: String = "",
                    @SerializedName("description")
@@ -86,7 +86,7 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("remarkName")
                    val remarkName: String? = null,
                    @SerializedName("accountStatus")
-                   val accountStatus: Int = 0,
+                   val accountStatus: Long = 0,
                    @SerializedName("avatarImgId")
                    val avatarImgId: Long = 0,
                    @SerializedName("defaultAvatar")
@@ -96,28 +96,28 @@ data class Profile(@SerializedName("detailDescription")
                    @SerializedName("backgroundImgIdStr")
                    val backgroundImgIdStr: String = "",
                    @SerializedName("province")
-                   val province: Int = 0,
+                   val province: Long = 0,
                    @SerializedName("nickname")
                    val nickname: String = "",
                    @SerializedName("djStatus")
-                   val djStatus: Int = 0,
+                   val djStatus: Long = 0,
                    @SerializedName("avatarUrl")
                    val avatarUrl: String = "",
                    @SerializedName("authStatus")
-                   val authStatus: Int = 0,
+                   val authStatus: Long = 0,
                    @SerializedName("vipType")
-                   val vipType: Int = 0,
+                   val vipType: Long = 0,
                    @SerializedName("followed")
                    val followed: Boolean = false,
                    @SerializedName("userId")
-                   val userId: Int = 0,
+                   val userId: Long = 0,
                    @SerializedName("mutual")
                    val mutual: Boolean = false,
                    @SerializedName("authority")
-                   val authority: Int = 0,
+                   val authority: Long = 0,
                    @SerializedName("backgroundImgId")
                    val backgroundImgId: Long = 0,
                    @SerializedName("userType")
-                   val userType: Int = 0)
+                   val userType: Long = 0)
